Move extensions from `PostRepository+Helpers.swift` to WordPressData
